# docs/swagger.yaml
basePath: /
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Deep QMS backend application serving the core HR, IT, and specialized
    logic domains.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DML API
  version: "1.0"
paths:
...
  /api/v1/departments:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of departments for the authenticated
        tenant.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: size
        type: integer
      - description: Search term (name/code)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated department data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List departments
      tags:
      - Organization
  /api/v1/departments/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific department by its ID.
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Department data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a department
      tags:
      - Organization
...
  /api/v1/job-titles:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of job titles for the authenticated
        tenant.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: size
        type: integer
      - description: Search term (name/code)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated job title data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List job titles
      tags:
      - Organization
  /api/v1/job-titles/{id}:
    get:
      consumes:
      - application/json
      description: Get job title by ID
      parameters:
      - description: Job Title ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job title data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Job Title
      tags:
      - organization
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"